pre-processor/requirements.txt
torch>=1.9.0
torchvision>=0.10.0
clip@git+https://github.com/openai/CLIP.git
Pillow>=8.3.1
requests>=2.26.0
pymongo>=3.12.0
numpy>=1.21.0
scikit-learn>=0.24.2
tqdm>=4.62.0
aiohttp>=3.8.0
python-dotenv>=0.19.0